move datasets from views to pages for consistency
