+++ frontend/Pages/KeywordSets/BaseKeywordSetAddView.cs
@@ -1,8 +1,8 @@
 using frontend.Models;
-using frontend.ViewModels;
+using frontend.ViewModels.KeywordSets;
 using Microsoft.AspNetCore.Components;
 
-namespace frontend.Pages;
+namespace frontend.Pages.KeywordSets;
 
 public class BaseKeywordSetAddView : ComponentBase
 {
