+++ frontend/Pages/KeywordSets/BaseKeywordSetEditView.cs
@@ -1,8 +1,7 @@
-using frontend.Models;
-using frontend.ViewModels;
+using frontend.ViewModels.KeywordSets;
 using Microsoft.AspNetCore.Components;
 
-namespace frontend.Pages;
+namespace frontend.Pages.KeywordSets;
 
 public class BaseKeywordSetEditView : ComponentBase
 {
