+++ frontend/Pages/KeywordSets/BaseKeywordSetListView.cs
@@ -1,8 +1,7 @@
-using frontend.ViewModels;
+using frontend.ViewModels.KeywordSets;
 using Microsoft.AspNetCore.Components;
-using Serilog;
 
-namespace frontend.Pages;
+namespace frontend.Pages.KeywordSets;
 
 public class BaseKeywordSetListView : ComponentBase
 {
