+++ frontend/ViewModels/KeywordSets/KeywordSetListViewModel.cs
@@ -2,7 +2,7 @@ using System.ComponentModel;
 using frontend.Models;
 using frontend.Services;
 
-namespace frontend.ViewModels;
+namespace frontend.ViewModels.KeywordSets;
 
 public interface IKeywordSetListViewModel
 {
