+++ frontend/_Imports.razor
@@ -14,3 +14,4 @@
 @using frontend.ViewModels
 @using frontend.ViewModels.Repositories
 @using frontend.ViewModels.Commits
+@using frontend.ViewModels.KeywordSets
